refactor: remove static subtitles and clean up file structure
